refactor(workspace): extract useAddDocuments hook
Move add-documents upload queue state and handlers into a dedicated hook.
